@@ -4,7 +4,9 @@
   "description": "React components to wrap Velocity animations",
   "main": "index.js",
   "scripts": {
-    "demo": "./node_modules/.bin/webpack-dev-server --progress --colors --content-base demo --hot --inline --config demo/webpack.config.js"
+    "compile": "babel src --out-dir .",
+    "demo": "./node_modules/.bin/webpack-dev-server --progress --colors --content-base demo --hot --inline --config demo/webpack.config.js",
+    "prepublish": "npm run compile"
   },
   "repository": {
     "type": "git",
@@ -23,22 +25,26 @@
   "homepage": "https://github.com/twitter-fabric/velocity-react#readme",
   "dependencies": {
     "lodash": "^3.10.1",
-    "react-addons-transition-group": "^0.14.0 || ^15.0.0",
+    "prop-types": "^15.5.8",
+    "react-transition-group": "^1.1.2",
     "velocity-animate": "^1.4.0"
   },
   "peerDependencies": {
-    "react": "^0.14.0 || ^15.0.0",
-    "react-dom": "^0.14.0 || ^15.0.0"
+    "react": "^0.14.0 || ^15.0.0 || ^16.0.0",
+    "react-dom": "^0.14.0 || ^15.0.0 || ^16.0.0"
   },
   "devDependencies": {
+    "babel-cli": "^6.24.1",
     "babel-core": "^6.21.0",
     "babel-loader": "^6.2.10",
+    "babel-plugin-transform-class-properties": "^6.24.1",
     "babel-plugin-transform-object-rest-spread": "^6.20.2",
-    "babel-preset-es2015": "^6.18.0",
+    "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.16.0",
+    "create-react-class": "^15.5.2",
     "css-loader": "^0.19.0",
-    "react": "^15.0.0",
-    "react-dom": "^15.0.0",
+    "react": "^15.5.0",
+    "react-dom": "^15.5.0",
     "react-hot-loader": "^1.2.9",
     "react-tween-state": "^0.1.3",
     "style-loader": "^0.12.3",
