@@ -1,30 +1,27 @@
 var React = require('react');
+var PropTypes = require('prop-types');
 var twemoji = require('twemoji');
 var s = require('underscore.string');
 
 // Component that uses the Twitter twemoji library to turn emoji-rich strings into spans with
 // <img> tags in them (for cross-browser/os compatibility).
-var EmojiSpan = React.createClass({
-  displayName: 'Emoji',
+class EmojiSpan extends React.Component {
+  static propTypes = {
+    size: PropTypes.number,
+    children: PropTypes.string,
+  };
 
-  propTypes: {
-    size: React.PropTypes.number,
-    children: React.PropTypes.string,
-  },
+  static defaultProps = {
+    size: 36,
+  };
 
-  getDefaultProps: function () {
-    return {
-      size: 36,
-    };
-  },
-
-  render: function () {
+  render() {
     let {children, size, ...attrs} = this.props;
 
     return (
       <span {...attrs} dangerouslySetInnerHTML={{__html: twemoji.parse(s.escapeHTML(children), {size: size})}} />
     );
-  },
-});
+  }
+}
 
 module.exports = EmojiSpan;
